+++ src/test/java/blackjack/domain/card/CardTypeTest.java
@@ -1,0 +1,15 @@
+package blackjack.domain.card;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class CardTypeTest {
+
+    @Test
+    @DisplayName("카드는 4가지 문양을 가질 수 있다")
+    void cardTypeTest() {
+        assertThat(CardType.values()).hasSize(4);
+    }
+}
