+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -1,0 +1,18 @@
+package blackjack.domain.card;
+
+import org.junit.jupiter.api.Test;
+
+class CardsTest {
+
+    @Test
+    void sum() {
+    }
+
+    @Test
+    void isBlackjack() {
+    }
+
+    @Test
+    void isBust() {
+    }
+}
