+++ src/test/java/blackjack/domain/deck/RandomCardStrategyTest.java
@@ -1,0 +1,10 @@
+package blackjack.domain.deck;
+
+import org.junit.jupiter.api.Test;
+
+class RandomCardStrategyTest {
+
+    @Test
+    void generateDeck() {
+    }
+}
