+++ src/test/java/blackjack/domain/gamer/PlayersTest.java
@@ -1,0 +1,10 @@
+package blackjack.domain.gamer;
+
+import org.junit.jupiter.api.Test;
+
+class PlayersTest {
+
+    @Test
+    void initialize() {
+    }
+}
