+++ src/test/java/blackjack/util/ListMergerTest.java
@@ -1,0 +1,10 @@
+package blackjack.util;
+
+import org.junit.jupiter.api.Test;
+
+class ListMergerTest {
+
+    @Test
+    void combine() {
+    }
+}
